requirements.txt
# native dependencies
Pillow==8.2.0
lxml==4.6.3
psycopg2==2.9.1
Django==3.2.4

# Other
amqp==5.0.6
beautifulsoup4==4.9.3
httplib2<0.19.2
hyperlink==21.0.0
idna>=2.5,<2.11
urllib3==1.26.5
Paver==1.3.4
python-slugify==5.0.2
decorator==5.0.9
celery==5.1.1
kombu==5.1.0
vine==5.0.0
tqdm==4.61.1
Deprecated==1.2.12
wrapt==1.12.1
jsonschema==3.2.0

# Django Apps
django-allauth==0.44.0
django-appconf==1.0.4
django-filter==2.4.0
django-imagekit==4.0.2
django-taggit==1.4.0
django-markdownify==0.9.0
django-mptt==0.12.0
django-modeltranslation>=0.11,<0.18.0
django-treebeard==4.5.1
django-guardian<2.4.1
django-downloadview==2.1.1
django-polymorphic==3.0.0
django-tastypie<0.15.0
django-tinymce==3.3.0
django-grappelli==2.15.1
django-uuid-upload-path==1.0.0
django-widget-tweaks==1.4.8
django-sequences==2.6
oauthlib==3.1.1
pyopenssl==20.0.1

# geopython dependencies
pyproj==3.1.0
OWSLib==0.24.1
pycsw==2.6.0
SQLAlchemy==1.4.18 # required by PyCSW
Shapely==1.7.1
mercantile==1.2.1
geoip2==4.2.0
numpy==1.19.*

# # Apps with packages provided in GeoNode's PPA on Launchpad.

# Django Apps
dj-database-url==0.5.0
dj-pagination==2.5.0
django-floppyforms<1.10.0
django-forms-bootstrap<=3.1.0
django-autocomplete-light==3.5.1
django-invitations<1.9.4
django-recaptcha==2.0.6

# REST
djangorestframework<3.12.0,>=3.8.0
djangorestframework-gis==0.17
djangorestframework-guardian==0.3.0
drf-extensions==0.7.0
drf-writable-nested==0.6.3
drf-spectacular==0.17.2
git+https://github.com/GeoNode/dynamic-rest.git@v2.0.1#egg=dynamic-rest
Markdown==3.3.4

pinax-notifications==6.0.0
pinax-ratings==4.0.0

# GeoNode org maintained apps.
# django-geonode-mapstore-client>=2.1.1
-e git+https://github.com/geosolutions-it/geonode-mapstore-client.git@nexus_master#egg=django_geonode_mapstore_client
geonode-avatar==5.0.7
geonode-oauth-toolkit==2.2.1
geonode-user-messages==2.0.2
geonode-announcements==2.0.2
geonode-django-activity-stream==0.10.0
gn-arcrest==10.5.4
geonode-dialogos==3.0.1
geoserver-restconfig==2.0.4.9
gn-gsimporter==2.0.1
gisdata==0.5.4

# haystack/elasticsearch
django-haystack==3.0
elasticsearch>=2.0.0,<8.0.0

# datetimepicker widget
django-bootstrap3-datetimepicker-2==2.8.2

# storage manager dependencies
django-storages==1.11.1
dropbox==11.9.0
google-cloud-storage==1.38.0
google-cloud-core==1.6.0
boto3==1.17.97

# Django Caches
python-memcached<=1.59

# Contribs
xmltodict<0.12.1
inflection>=0.4.0
jdcal==1.4.1
mock<5.0.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
timeout-decorator==0.5.0

# required by monitoring
psutil==5.8.0
django-cors-headers==3.7.0
user-agents
xmljson
django-ipware<3.1
# no version here, use latest one with fresh data
pycountry

# production
uWSGI==2.0.19.1
gunicorn==20.1.0
ipython==7.24.1
docker==4.4.4
invoke==1.5.0

# tests
coverage==5.5
requests-toolbelt==0.9.1
flake8==3.9.2
pytest==6.2.4
pytest-bdd==4.0.2
splinter==0.14.0
pytest-splinter==3.3.1
pytest-django==4.4.0
setuptools==57.0.0
pip==21.1.2
Twisted==21.2.0
pixelmatch==0.2.3
factory-boy==3.2.0
flaky==3.7.0
selenium==3.141.0
selenium-requests==1.3
webdriver_manager==3.4.2
